bpf: move remaining node-port handling into header
We need this later on for extending bpf_overlay as well, therefore
move it out of bpf_netdev.
